Fix missing deduplication in constraint tracking
Prior to this fix the same dependency state could lead to duplicate
outgoing edges which would cause an issue when the node targeted was
removed because it was deselected.
The code now uses a datastructure that will perform the de duplication
and it no longer registers entries in some code paths.

Fixes #13551